requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beartype==0.16.4
beautifulsoup4==4.12.2
bleach==6.1.0
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.3.2
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.6.0
ema-pytorch==0.3.1
exceptiongroup==1.1.3
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.104.1
fastjsonschema==2.19.1
ffmpy==0.3.1
filelock @ file:///croot/filelock_1672387128942/work
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.40
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gradio==4.7.1
gradio_client==0.7.0
grpcio==1.60.0
h11==0.14.0
h5py==3.9.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.18.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.5
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib-resources==6.1.0
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1694181358621/work
ipywidgets==8.1.1
isoduration==20.11.0
jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.3.2
json5==0.9.22
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.1
jupyter-lsp==2.2.4
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1698937308754/work
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.4
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
kiwisolver==1.4.5
lazy_loader==0.3
lightning==2.1.1
lightning-utilities==0.9.0
lmdb==1.3.0
lpips==0.1.4
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.8.0
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.4
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.2
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx @ file:///croot/networkx_1690561992265/work
notebook==7.1.1
notebook_shim==0.2.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.6.0.66
orjson==3.9.10
overrides==7.7.0
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.1.1
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.3.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
POT==0.9.1
prometheus_client==0.20.0
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments @ file:///croot/pygments_1684279966437/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-fid==0.3.0
pytorch-lightning==2.1.1
pytorch-ssim==0.1
pytz==2023.3.post1
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq @ file:///croot/pyzmq_1705605076900/work
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.0
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.9.3
seaborn==0.13.0
semantic-version==2.10.0
Send2Trash==1.8.2
sentry-sdk==1.32.0
setproctitle==1.3.3
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
sk-video==1.1.10
sklearn==0.0.post10
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.27.0
sympy @ file:///croot/sympy_1668202399572/work
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.5.1
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.6.12
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==1.13.0
torchaudio==2.1.0
torchmetrics==1.2.0
torchstat==0.0.7
torchsummaryX==1.3.0
torchvision==0.14.0
tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.66.1
traitlets==5.11.2
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.20240311
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1686163155763/work
uvicorn==0.24.0.post1
wandb==0.16.0
warmup-scheduler==0.3
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.9
yacs==0.1.8
yarl==1.9.2
zipp @ file:///croot/zipp_1704206909481/work
